updates for constraint groups, setup to support more flexible bool queries
